aes-gcm: add example for doing AES-GCM calls
Add sample code that shows how to wrap GlobalPlatform Internal API
calls to be able to do AES-GCM encryption/decryption.

The normal world code only kicks of the test in the TA, so it is the TA
itself that is doing all work and it leverages test vectors that are
stored locally withing the TA itself. However, only minor changes are
needed if someone would like to provide buffers from normal world
instead.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>